Split out util.c
