+++ include/novas.h
@@ -1258,9 +1258,11 @@ double novas_inv_refract(RefractionModel model, double jd_tt, const on_surface *
 // ---------------------- Added in 1.2.0 -------------------------
 int make_redshifted_object(const char *name, double ra, double dec, double z, object *source);
 
-double novas_v2z(double vel);
-
 double novas_z2v(double z);
+
+
+// in util.c
+double novas_v2z(double vel);
 
 double grav_redshift(double M_kg, double r_m);
